backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# Azure Document Intelligence
azure-ai-formrecognizer==3.3.3
azure-core==1.31.0

# Claude API
anthropic>=0.43.0

# Image processing
opencv-python-headless==4.8.1.78
Pillow==10.4.0
pdf2image==1.17.0
img2pdf==0.5.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
faker==30.8.2
gender-guesser==0.4.0

# CORS
python-jose[cryptography]==3.3.0
